Add files via upload
Here is just some background if you will on how the Incident Command System works currently and how "people" accountability is limited to "overhead" personnel only.  I put this together not to say what we all already know but more to give you an idea of what everyone in the response industry knows as the standard according to the National Response Framework (NRF) and the National Incident Management System (NIMS).  It may be useful as we go down this road and get through the code configuration and it takes shape; first responder's (especially the leadership) are very familiar with this system and like it (for the most part) and maybe we can glean some of the info from the docs as standard inputs into the app.  Maybe the app itself is shaped like our "T" card, anyway maybe I'm way off here.