Update and rename mike_infotech_questions.txt to mike_ITadmin_questions.txt
For what it's worth:
Sharpen the focus of these interviews and use these to understand the problem before trying to quantify it. These people do not have a lot of time, nor do we, so the interviewer should be up front about what we want otherwise an interviewee will hold back wondering what your real agenda is. I've had decades of training and experience interviewing about complex education policies using multi-method research design, triangulating sources in order to zero in on the best target for adoption. So given that you want to propose that for Phase II, it's something you may want to consider. Also a strategy I've had success with, particularly for multi-year projects, is to confirm a theory among respondents, then once you have established trust you can always go back and ask for specific numbers. For example, "You mentioned that you've seen this problem frequently, is that once a month, a week? You've accomplished two things -- collected additional data and captured a source to go back to. I say this because I sense you feel you need to make the case and quantify it too and that is probably not realistic give then size and diversity of the emergency responder community. Your results will be skewed, inadequate, and potentially misleading. I would make that part of your R & D budget, put aside a small item line for that.